LCH-6356: Fixing contenthubloggingclient.
